+++ src/main/java/org/hh/heritagehunters/domain/post/service/CommentService.java
@@ -21,15 +21,11 @@ public class CommentService {
   private final CommentRepository commentRepository;
 
   public Comment add(Long postId, User user, CommentCreateRequestDto dto) {
-
     Post post = postRepository.findById(postId)
         .orElseThrow(() -> new NotFoundException(ErrorCode.POST_NOT_FOUND));
-
     Comment comment = Comment.create(user, post, dto.getContent());
     commentRepository.save(comment);
-
     post.syncCommentCount();
-
     return comment;
   }
 }
