+++ src/main/resources/static/features/post/css/modal-common.css
