+++ src/main/resources/templates/features/post/post_detail.html
@@ -6,7 +6,7 @@
       th:insert="~{fragments/head :: common-head('Postdetail - ' + ${post.id})}"></th:block>
   <meta name="_csrf" th:content="${_csrf.token}"/>
   <meta name="_csrf_header" th:content="${_csrf.headerName}"/>
-  <link rel="stylesheet" th:href="@{/features/post/css/modal_common.css}">
+  <link rel="stylesheet" th:href="@{/features/post/css/modal-common.css}">
   <link rel="stylesheet" th:href="@{/features/post/css/post_write.css}">
   <link rel="stylesheet" th:href="@{/features/post/css/post_detail.css}">
 </head>
