+++ src/main/resources/templates/features/post/post_write.html
@@ -3,7 +3,7 @@
       xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
 <head>
   <th:block th:insert="~{fragments/head :: common-head('Postwrite')}"></th:block>
-  <link rel="stylesheet" th:href="@{/features/post/css/modal_common.css}">
+  <link rel="stylesheet" th:href="@{/features/post/css/modal-common.css}">
   <link rel="stylesheet" th:href="@{/features/post/css/post_write.css}">
 </head>
 <body>
