+++ src/main/resources/templates/features/profile/profile_page.html
@@ -10,7 +10,6 @@
     <!-- CSS -->
     <link rel="stylesheet" th:href="@{/features/profile/css/profile_page.css}" />
     <link rel="stylesheet" th:href="@{/features/profile/css/modal.css}" />
-
   </head>
   <body>
     <!--header-->
